+++ src/main/java/softeer2nd/chess/ChessGame.java
@@ -1,11 +1,15 @@
 package softeer2nd.chess;
 
+import softeer2nd.chess.pieces.Color;
 import softeer2nd.chess.pieces.Piece;
 import softeer2nd.chess.pieces.Type;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
+import static softeer2nd.chess.Board.COl;
 import static softeer2nd.chess.pieces.Piece.createBlank;
 
 public class ChessGame {
@@ -16,6 +20,9 @@ public class ChessGame {
         this.board = board;
     }
 
+    /**
+     * 기물 이동
+     */
     public void move(String sourcePosition, String targetPosition) {
         Piece findPiece = findPiece(sourcePosition);
 
@@ -47,4 +54,45 @@ public class ChessGame {
         return createBlank(findPosition);
     }
 
+    /**
+     * 점수 계산
+     */
+    public double calculatePoint(Color color) {
+        double point = 0;
+
+        if(color.equals(Color.BLACK)) {
+            point = getPoint(point, board.getBlackPieces());
+        } else if(color.equals(Color.WHITE)){
+            point = getPoint(point, board.getWhitePieces());
+        }
+
+        return point;
+    }
+
+    private double getPoint(double point, Map<Type, List<Piece>> colorPieces) {
+        for (Type type : Type.values()) {
+            List<Piece> pieces = colorPieces.getOrDefault(type, new ArrayList<>());
+
+            if(type.equals(Type.PAWN)) {
+                HashMap<Integer, Integer> colSet = new HashMap<>();
+
+                for (Piece piece : pieces) {
+                    int col = piece.getPosition().getCol();
+                    colSet.put(col, colSet.getOrDefault(col, 0) + 1);
+                }
+
+                int sameColPawnCnt = 0;
+                for (int i = 0; i < COl; i++) {
+                    double cnt = colSet.getOrDefault(i, 0);
+                    if(cnt > 1) sameColPawnCnt += cnt;
+                }
+                point -= sameColPawnCnt / 2;
+
+            }
+
+            point += type.getDefaultPoint() * pieces.size();
+        }
+
+        return point;
+    }
 }
